clean data processing pipeline and repo
